# backend/auth-service/docs/swagger.yaml
basePath: /api
definitions:
  handlers.ApproveOrganizerRequest:
    properties:
      user_id:
        type: integer
    required:
    - user_id
    type: object
  handlers.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handlers.ResetPasswordRequest:
    properties:
      code:
        type: string
      email:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - code
    - email
    - new_password
    type: object
  handlers.VerifyEmailRequest:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
  models.ApprovalStatus:
    enum:
    - pending
    - approved
    - rejected
    type: string
    x-enum-varnames:
    - ApprovalStatusPending
    - ApprovalStatusApproved
    - ApprovalStatusRejected
  models.Role:
    enum:
    - admin
    - organizer
    - user
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleOrganizer
    - RoleUser
  models.User:
    properties:
      approval_status:
        allOf:
        - $ref: '#/definitions/models.ApprovalStatus'
        description: Default approved for users
      created_at:
        type: string
      deleted_at:
        type: string
      document_path:
        description: For organizers
        type: string
      email:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      role:
        $ref: '#/definitions/models.Role'
      updated_at:
        type: string
    type: object
host: localhost:3001
info:
  contact: {}
  description: This is the authentication service.
  title: Auth Service API
  version: "1.0"
paths:
  /auth/admin/approve:
    post:
      consumes:
      - application/json
      description: Approve a pending organizer account (Admin only)
      parameters:
      - description: Approve Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.ApproveOrganizerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Approve Organizer
      tags:
      - admin
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Request password reset code
      parameters:
      - description: Forgot Password Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Forgot Password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password
      parameters:
      - description: Login Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get current user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Profile
      tags:
      - auth
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Get new access token using refresh token
      parameters:
      - description: Refresh Token Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Refresh Token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - multipart/form-data
      description: Register a new user with email, password, and role. Organizers
        must upload a document.
      parameters:
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      - description: Role (user or organizer)
        in: formData
        name: role
        required: true
        type: string
      - description: Document (required for organizer)
        in: formData
        name: document
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password with code
      parameters:
      - description: Reset Password Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Reset Password
      tags:
      - auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Verify user email with code
      parameters:
      - description: Verify Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Verify Email
      tags:
      - auth
swagger: "2.0"